refactor(get-help): modularize components and add unique IDs
Refactored the Get Help feature by breaking it into modular components and improving maintainability.

- Modularized `Header`, `InputArea`, `InputField`, `Message`, `MessageList`, and `SuggestionButtons` as separate components
- Introduced unique IDs for `LiveFeedbackMessage` using `uniqueId` for better tracking
- Enhanced typings with a new `EditorRef` interface
- Updated reducer logic to accommodate the new `id` field in messages
- Simplified and improved readability of the `GetHelpPage` component
